+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# 0.0.6
-Re-write IteratorContainerList and IteratorContainerSet
+# 0.0.7
+Optimize the stream & spliterator generation of IteratorContainerList.
+
+Never validate container if the given iterator is empty.
+
+Fix isEmpty checking logic being incorrect.